Fix bug in sequenceNode.removeGenome removes index
The method calls remove(int) on an Integer list, meaning it will resolve
to remove(index) rather than remove(object).

Also amended the tests to catch this mistake.

+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/SequenceNode.java
@@ -181,7 +181,7 @@ public class SequenceNode extends AbstractNode {
   public void removeGenome(int genome) {
     assert genomes.contains(
         genome) : "Removing non-existent genome: " + genome + ". NodeID: " + this.getId();
-    genomes.remove(genome);
+    genomes.remove((Integer) genome);
   }
 
   @Override
